Deploy as apps/requirements.txt
streamlit>=1.46.0
openai>=1.90.0
python-dotenv>=1.1.0
requests>=2.32.4
selenium>=4.33.0
webdriver-manager>=4.0.2
beautifulsoup4>=4.12.0
lxml>=4.9.0
urllib3>=2.0.0 